Pull out code into quick-bet.tsx
